separate expand into multiple modules
