+++ leptos-mview-core/src/expand.rs
@@ -8,18 +8,52 @@ use std::collections::HashMap;
 use proc_macro2::{Span, TokenStream};
 use proc_macro_error::emit_error;
 use quote::{quote, quote_spanned, ToTokens};
-use syn::spanned::Spanned;
 
 use crate::ast::{
-    attribute::{
-        directive::Directive,
-        kv::KvAttr,
-        parsing::KebabIdentOrStr,
-        selector::{SelectorShorthand, SelectorShorthands},
-        spread_attrs::SpreadAttr,
-    },
+    attribute::{directive::Directive, parsing::KebabIdentOrStr, selector::SelectorShorthand},
     Attr, Element, KebabIdent, NodeChild, Tag,
 };
+
+/// Functions for specific parts of an element's expansion.
+mod subroutines;
+use subroutines::*;
+/// Small helper functions for converting types or emitting errors.
+mod utils;
+use utils::*;
+
+/// Converts the children into a `leptos::Fragment::lazy()` token stream.
+///
+/// Example:
+/// ```ignore
+/// "a"
+/// {var}
+/// "b"
+/// ```
+///
+/// Should expand to:
+/// ```ignore
+/// Fragment::lazy(|| {
+///     [
+///         {"a"}.into_view(),
+///         {var}.into_view(),
+///         {"b"}.into_view(),
+///     ].to_vec()
+/// })
+/// ```
+
+// used in the root or for component children
+pub(crate) fn children_fragment_tokens<'a>(
+    children: impl Iterator<Item = &'a NodeChild>,
+    span: Span,
+) -> TokenStream {
+    quote_spanned! { span=>
+        ::leptos::Fragment::lazy(|| {
+            ::std::vec![
+                #(  ::leptos::IntoView::into_view(#children) ),*
+            ]
+        })
+    }
+}
 
 /// Converts an xml (like html, svg or math) element to tokens.
 ///
@@ -81,7 +115,7 @@ pub fn xml_to_tokens(element: &Element) -> Option<TokenStream> {
 
     let children = element
         .children()
-        .map(|children| child_methods_tokens(children.element_children()));
+        .map(|children| xml_child_methods_tokens(children.element_children()));
 
     Some(quote! {
         #tag_path
@@ -91,100 +125,6 @@ pub fn xml_to_tokens(element: &Element) -> Option<TokenStream> {
             #spread_attrs
             #children
     })
-}
-
-/// Converts element class/id selector shorthands into a series of `.classes`
-/// and `.id` calls.
-fn xml_selectors_tokens(selectors: &SelectorShorthands) -> TokenStream {
-    let (classes, ids): (Vec<_>, Vec<_>) = selectors
-        .iter()
-        .partition(|sel| matches!(sel, SelectorShorthand::Class { .. }));
-
-    let class_methods = classes.iter().map(|class| {
-        let method = syn::Ident::new("class", class.prefix().span());
-        let class_name = class.ident().to_str_colored();
-        quote! { .#method(#class_name, true) }
-    });
-
-    let id_methods = ids.iter().map(|id| {
-        let method = syn::Ident::new("id", id.prefix().span());
-        let id_name = id.ident().to_str_colored();
-        quote! { .#method(#id_name) }
-    });
-
-    quote! { #(#class_methods)* #(#id_methods)* }
-}
-
-fn xml_kv_attribute_tokens(attr: &KvAttr) -> TokenStream {
-    let key = attr.key();
-    let value = attr.value();
-    // special cases
-    if key.repr() == "ref" {
-        let node_ref = syn::Ident::new("node_ref", key.span());
-        quote! { .#node_ref(#value) }
-    } else {
-        quote! { .attr(#key, #value) }
-    }
-}
-
-fn emit_error_if_modifier(m: Option<&syn::Ident>) {
-    if let Some(modifier) = m {
-        emit_error!(
-            modifier.span(),
-            "modifiers are only supported on `on:` directives"
-        );
-    }
-}
-
-fn xml_directive_tokens(directive: &Directive) -> TokenStream {
-    let Directive {
-        dir,
-        key,
-        modifier,
-        value,
-    } = directive;
-
-    match dir.to_string().as_str() {
-        "class" | "style" | "prop" => {
-            let key = key.to_lit_str();
-            emit_error_if_modifier(modifier.as_ref());
-            quote! { .#dir(#key, #value) }
-        }
-        "on" => event_listener_tokens(directive),
-        "use" => use_directive_to_method(directive),
-        "attr" | "clone" => {
-            emit_error!(dir.span(), "`{}:` is not supported on elements", dir);
-            quote! {}
-        }
-        _ => {
-            emit_error!(dir.span(), "unknown directive");
-            quote! {}
-        }
-    }
-}
-
-fn xml_spread_tokens(attr: &SpreadAttr) -> TokenStream {
-    let (dotdot, expr) = (attr.dotdot(), attr.expr());
-    let attrs = syn::Ident::new("attrs", dotdot.span());
-    quote! {
-        .#attrs(#expr)
-    }
-}
-
-/// Converts the children to a series of `.child` calls.
-///
-/// # Example
-/// ```ignore
-/// div { "a" {var} "b" }
-/// ```
-/// Expands to:
-/// ```ignore
-/// div().child("a").child({var}).child("b")
-/// ```
-pub fn child_methods_tokens<'a>(children: impl Iterator<Item = &'a NodeChild>) -> TokenStream {
-    quote! {
-        #( .child(#children) )*
-    }
 }
 
 /// Transforms a component into a `TokenStream` of a leptos component view.
@@ -317,7 +257,7 @@ pub fn component_to_tokens<const IS_SLOT: bool>(element: &Element) -> Option<Tok
         .children()
         .map(|children| slots_to_tokens(children.slot_children()));
 
-    let dyn_attrs = dyn_attrs_to_methods(&dyn_attrs);
+    let dyn_attrs = component_dyn_attrs_to_methods(&dyn_attrs);
     let use_directives = use_directives.into_iter().map(use_directive_to_method);
     let dyn_classes = component_classes_to_method(dyn_classes);
     let selector_ids = component_ids_to_method(selector_ids);
@@ -357,307 +297,6 @@ pub fn component_to_tokens<const IS_SLOT: bool>(element: &Element) -> Option<Tok
             .into_view()
             #(#use_directives)*
             #event_listeners
-        })
-    }
-}
-
-fn component_kv_attribute_tokens(attr: &KvAttr) -> TokenStream {
-    let (key, value) = (attr.key().to_snake_ident(), attr.value());
-    quote_spanned! { attr.span()=> .#key(#value) }
-}
-
-fn event_listener_tokens(dir: &Directive) -> TokenStream {
-    let Directive {
-        dir,
-        key,
-        modifier,
-        value,
-    } = dir;
-    if dir != "on" {
-        panic!("directive should be `on:`");
-    };
-
-    let ev_name = match key {
-        KebabIdentOrStr::KebabIdent(ident) => ident.to_snake_ident(),
-        KebabIdentOrStr::Str(s) => {
-            emit_error!(s.span(), "event type must be an identifier");
-            syn::Ident::new("invalid_event", s.span())
-        }
-    };
-
-    let event = if let Some(modifier) = modifier {
-        if modifier.to_string() == "undelegated" {
-            quote! { ::leptos::ev::#modifier(::leptos::ev::#ev_name) }
-        } else {
-            emit_error!(
-                modifier.span(), "unknown modifier";
-                help = ":undelegated is the only known modifier"
-            );
-            quote! { ::leptos::ev::#ev_name }
-        }
-    } else {
-        quote! { ::leptos::ev::#ev_name }
-    };
-    quote! { .#dir(#event, #value) }
-}
-
-/// Expands to a `let` statement `let to_clone = to_clone.clone();`.
-fn component_clone_tokens(dir: &Directive) -> TokenStream {
-    let to_clone = dir.key.to_ident_or_emit();
-    emit_error_if_modifier(dir.modifier.as_ref());
-    quote! { let #to_clone = #to_clone.clone(); }
-}
-
-/// Converts children to tokens for use by components.
-///
-/// The expansion is generally:
-///
-/// If there are no closure arguments,
-/// ```ignore
-/// .children({
-///     // any clones
-///     let clone = clone.clone();
-///     // the children themself
-///     leptos::ToChildren::to_children(move || {
-///         leptos::Fragment::lazy(|| {
-///             [
-///                 child1.into_view(),
-///                 child2.into_view(),
-///             ].to_vec()
-///         })
-///     })
-/// })
-/// ```
-///
-/// If there are closure arguments,
-/// ```ignore
-/// .children({
-///     // any clones
-///     let clone = clone.clone();
-///     // the children
-///     move |args| leptos::Fragment::lazy(|| {
-///         [
-///             child1.into_view(),
-///             child2.into_view(),
-///         ].to_vec()
-///     })
-/// })
-/// ```
-fn component_children_tokens<'a>(
-    children: impl Iterator<Item = &'a NodeChild>,
-    args: Option<&TokenStream>,
-    clones: &TokenStream,
-) -> TokenStream {
-    let mut children = children.peekable();
-    let child_span = children
-        .peek()
-        // not sure why `child.span()` is calling `syn::spanned::Spanned` instead
-        .map_or_else(Span::call_site, |child| (*child).span());
-
-    // span call site if there are no args so that the children don't get all the
-    // `std` `vec!` etc docs.
-    let children_fragment =
-        children_fragment_tokens(children, args.map_or(Span::call_site(), Spanned::span));
-
-    // children with arguments take a `Fn(T) -> impl IntoView`
-    // normal children (`Children`, `ChildrenFn`, ...) take
-    // `ToChildren::to_children`
-    let wrapped_fragment = if let Some(args) = args {
-        // `args` includes the pipes
-        quote_spanned!(args.span()=> move #args #children_fragment)
-    } else {
-        // this span is required for slots that take `Callback<T, View>` but have been
-        // given a regular `ChildrenFn` instead.
-        let closure = quote_spanned!(child_span=> move || #children_fragment);
-        quote! {
-            ::leptos::ToChildren::to_children(#closure)
-        }
-    };
-
-    let children_method = quote_spanned!(child_span=> children);
-
-    quote! {
-        .#children_method({
-            #clones
-            #wrapped_fragment
-        })
-    }
-}
-
-fn dyn_attrs_to_methods(dyn_attrs: &[&Directive]) -> Option<TokenStream> {
-    // expand dyn attrs to the method if any exist
-    if dyn_attrs.is_empty() {
-        return None;
-    };
-
-    let dyn_attrs_method = syn::Ident::new("dyn_attrs", dyn_attrs[0].dir.span());
-
-    let (keys, values): (Vec<_>, Vec<_>) = dyn_attrs
-        .into_iter()
-        .map(|a| (a.key.to_lit_str(), &a.value))
-        .unzip();
-
-    Some(quote! {
-        .#dyn_attrs_method(
-            ::std::vec![
-                #( (#keys, ::leptos::IntoAttribute::into_attribute(#values)) ),*
-            ]
-        )
-    })
-}
-
-// special attributes on components that add to a special set of props //
-
-/// Adds potentially reactive classes to the `class` attribute of a component.
-///
-/// If no classes are reactive, a static string will be passed in. Otherwise,
-/// the string is constructed and updated at runtime, which may have performance
-/// drawbacks as the entire prop is updated if one signal changes.
-///
-/// The intended use is as follows:
-/// ```ignore
-/// #[component]
-/// fn TakesClasses(#[prop(optional, into)] class: TextProp) -> impl IntoView {}
-///
-/// let signal = RwSignal::new(true);
-///
-/// mview! {
-///     TakesClasses.class-1.another-class class:reactive={signal};
-/// }
-/// ```
-///
-/// For now, what is passed in to `{signal}` must be something that impls `Fn()
-/// -> bool`, it cannot just be a `bool`.
-fn component_classes_to_method(
-    classes: Vec<(KebabIdentOrStr, Option<TokenStream>)>,
-) -> Option<TokenStream> {
-    if classes.is_empty() {
-        return None;
-    };
-
-    let first_span = classes[0].0.span();
-
-    // if there are no reactive classes, just create the string now
-    // add `||` to reject `class:thing={true}`
-    if classes.iter().all(|(_, signal)| signal.is_none()) {
-        let string = classes
-            .into_iter()
-            .map(|(class, _)| class.to_lit_str().value())
-            .collect::<Vec<_>>()
-            .join(" ");
-        Some(quote_spanned!(first_span=> .class(#string)))
-    } else {
-        // there are reactive classes: need to construct it at runtime
-
-        // TODO: is there a way to accept both `bool` and `Fn() -> bool`?
-        // maybe `leptos::Class`?
-
-        let classes_array = classes.into_iter().map(|(class, signal)| {
-            // add extra bracket to make sure the closure is called
-            let signal_called = quote_spanned! { signal.span()=> (#signal)() };
-            let class = class.to_lit_str();
-
-            // use fully qualified path so that error says 'incorrect type' instead of
-            // 'method `then_some` not found'
-            quote_spanned! { signal_called.span()=>
-                ::std::primitive::bool::then_some(#signal_called, #class)
-            }
-        });
-        let classes_array = quote_spanned!(first_span=> [#(#classes_array),*]);
-        let contents = quote_spanned! { first_span=>
-            #classes_array
-                .iter()
-                .flatten() // remove None
-                .cloned() // turn &&str to &str
-                .collect::<Vec<&str>>()
-                .join(" ")
-        };
-
-        // span to the first item
-        Some(quote_spanned! { first_span=>
-            .class(move || #contents)
-        })
-    }
-}
-
-/// Adds a list of strings to the `id` prop of a component.
-///
-/// IDs should not be changed reactively, so it is not supported.
-fn component_ids_to_method(ids: Vec<syn::LitStr>) -> Option<TokenStream> {
-    if ids.is_empty() {
-        return None;
-    };
-
-    let first_span = ids[0].span();
-    // ids are not reactive, so just give one big string
-    let ids = ids
-        .into_iter()
-        .map(|id| id.value())
-        .collect::<Vec<_>>()
-        .join(" ");
-
-    Some(quote_spanned!(first_span=> .id(#ids)))
-}
-
-/// Converts a `use:directive={value}` to a method.
-///
-/// The expansion for components and xml elements are the same.
-///
-/// ```text
-/// use:d => .directive(d, ().into())
-/// use:d={some_value} => .directive(d, some_value.into())
-/// ```
-///
-/// **Panics** if the provided directive is not `use:`.
-fn use_directive_to_method(u: &Directive) -> TokenStream {
-    let Directive {
-        dir: use_token,
-        key,
-        modifier,
-        value,
-    } = u;
-    if use_token != "use" {
-        panic!("directive should be `use:`")
-    };
-    let directive_fn = key.to_ident_or_emit();
-    emit_error_if_modifier(modifier.as_ref());
-
-    let directive = syn::Ident::new("directive", use_token.span());
-    let value = value.as_ref().map_or_else(
-        || quote_spanned! {directive_fn.span()=> ().into() },
-        |val| quote! { ::std::convert::Into::into(#val) },
-    );
-    quote! { .#directive(#directive_fn, #value) }
-}
-
-/// Converts the children into a `leptos::Fragment::lazy()` token stream.
-///
-/// Example:
-/// ```ignore
-/// "a"
-/// {var}
-/// "b"
-/// ```
-///
-/// Should expand to:
-/// ```ignore
-/// Fragment::lazy(|| {
-///     [
-///         {"a"}.into_view(),
-///         {var}.into_view(),
-///         {"b"}.into_view(),
-///     ].to_vec()
-/// })
-/// ```
-pub fn children_fragment_tokens<'a>(
-    children: impl Iterator<Item = &'a NodeChild>,
-    span: Span,
-) -> TokenStream {
-    quote_spanned! { span=>
-        ::leptos::Fragment::lazy(|| {
-            ::std::vec![
-                #(  ::leptos::IntoView::into_view(#children) ),*
-            ]
         })
     }
 }
@@ -708,7 +347,7 @@ fn slots_to_tokens<'a>(children: impl Iterator<Item = &'a Element>) -> TokenStre
         .into_iter()
         .map(|(slot_name, slot_tokens)| {
             let method = syn::Ident::new_raw(
-                &upper_camel_to_snake_case(slot_name.repr()),
+                &utils::upper_camel_to_snake_case(slot_name.repr()),
                 slot_name.span(),
             );
 
@@ -726,22 +365,4 @@ fn slots_to_tokens<'a>(children: impl Iterator<Item = &'a Element>) -> TokenStre
             }
         })
         .collect()
-}
-
-#[allow(clippy::doc_markdown)]
-// just doing a manual implementation as theres only one need for this (slots).
-// Use the `paste` crate if more are needed in the future.
-/// `ident` must be an UpperCamelCase word with only ascii word characters.
-fn upper_camel_to_snake_case(ident: &str) -> String {
-    let mut new = String::with_capacity(ident.len());
-    // all characters should be ascii
-    for char in ident.chars() {
-        // skip the first `_`.
-        if char.is_ascii_uppercase() && !new.is_empty() {
-            new.push('_');
-        };
-        new.push(char.to_ascii_lowercase());
-    }
-
-    new
 }
